requirements.txt
asgiref==3.11.1
colorama==0.4.6
Django==5.1.4
django-environ==0.11.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.0
factory-boy==3.3.1
Faker==40.13.0
iniconfig==2.3.0
mysqlclient==2.2.8
packaging==26.0
pluggy==1.6.0
pytest==8.3.4
pytest-django==4.9.0
sqlparse==0.5.5
tzdata==2026.1
django-extensions==3.2.3
bcrypt==4.2.0
PyMySQL==1.1.2
cryptography
Pygments==2.19.2
PyJWT==2.12.1
